+++ changelog.md
@@ -10,9 +10,11 @@ This project follows semVer, where:
  -->
 
 ### 7.10.0 [Apr 2025]
+- **[fix]** -  capitalization of "Etc/Utc" #446
 - **[new]** - `.isoFull()` method
 - **[new]** - `spacetime.fromEpochSeconds()` method #444
 - **[fix]** - getter form of `.epochSeconds()` 
+- **[update]** - dependencies
 
 ### 7.9.0 [Apr 2025]
 - **[new]** - `.format('sql')` #437